Added State Tracking and OnDataReceivedListener
This change adds some interfaces that are needed to support
my changes to the factory diagnostic app.

New interfaces are in I2CState.kt while they are implemented
in `SHT40Sensor`, `AS7343Sensor` and `TCA9548Multiplexer`

This enables more cleanly track the state of sensors
and expose them in the android ui.